+++ CHANGELOG.md
@@ -49,6 +49,11 @@ To install release candidates run `[sudo] gem install cocoapods --pre`
   [Whirlwind](https://github.com/Whirlwind)
   [#7878](https://github.com/CocoaPods/CocoaPods/pull/7878)
 
+* Set the path of the Pods group in the user project 
+  [Whirlwind](https://github.com/Whirlwind)
+  [#7886](https://github.com/CocoaPods/CocoaPods/pull/7886)
+  [#6194](https://github.com/CocoaPods/CocoaPods/issues/6194)
+
 * Add a `--deployment` flag to `pod install` that errors if there are any
   changes to the Podfile or Lockfile.  
   [Samuel Giddins](https://github.com/segiddins)